+++ internetarchive/cli/ia.py
@@ -100,6 +100,10 @@ def main():
                         action="store",
                         help=("host to connect to "
                               "(doesn't work for requests made to s3.us.archive.org)"))
+    parser.add_argument("--user-agent",
+                        action="store",
+                        metavar="STRING",
+                        help="custom User-Agent string for all requests")
 
     subparsers = parser.add_subparsers(title="commands",
                                        dest="command",
@@ -137,6 +141,11 @@ def main():
             config["general"]["host"] = args.host
         else:
             config["general"] = {"host": args.host}
+    if args.user_agent:
+        if config.get("general"):
+            config["general"]["user_agent"] = args.user_agent
+        else:
+            config["general"] = {"user_agent": args.user_agent}
 
     args.session = get_session(config_file=args.config_file,
                                config=config,